# server/docs/swagger.yaml
definitions:
...
  responses.Error:
    properties:
      error:
        type: object
    required:
    - error
    type: object
...
host: localhost:3002/api
info:
  contact: {}
  description: This is the API for Schej.it!
  license: {}
  title: Schej.it API
  version: "1.0"
paths:
...
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: Signs user in and sets the access token session variable
      parameters:
      - description: Object containing the Google authorization code, scope, calendar
          type, and the user's timezone offset
        in: body
        name: payload
        required: true
        schema:
          allOf:
          - type: object
          - properties:
              calendarType:
                type: string
              code:
                type: string
              scope:
                type: string
              timezoneOffset:
                type: integer
            type: object
      produces:
      - application/json
      responses:
        "200": {}
      summary: Signs user in
      tags:
      - auth
  /auth/sign-in-mobile:
    post:
      consumes:
      - application/json
      description: Signs user in and sets the access token session variable
      parameters:
      - description: Object containing the Google authorization code, calendar type,
          and the user's timezone offset
        in: body
        name: payload
        required: true
        schema:
          allOf:
          - type: object
          - properties:
              accessToken:
                type: string
              calendarType:
                type: string
              expiresIn:
                type: integer
              idToken:
                type: string
              refreshToken:
                type: string
              scope:
                type: string
              timezoneOffset:
                type: integer
              tokenOrigin:
                type: string
            type: object
      produces:
      - application/json
      responses:
        "200": {}
      summary: Signs user in from mobile
      tags:
      - auth
...
  /auth/status:
    get:
      description: Returns a 401 error if user is not signed in, 200 if they are
      responses:
        "200": {}
        "401":
          description: Error object
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Gets whether the user is signed in or not
      tags:
      - auth
...
swagger: "2.0"